+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to npad will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.32.4] - unreleased
+## [0.32.4] - 2026-09-17
 
 ### 🐛 Fixed
 - **Installer: the "Launch npad" boxes on the Finished page had the same